# coalgov-ai/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoalGov AI API",
    "description": "Mine governance pilot: inspections, compliance, production approvals, grievances and offline field reporting.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/intelligence": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Intelligence",
        "operationId": "intelligence_api_intelligence_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sensors": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Sensors",
        "operationId": "sensors_api_sensors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/telemetry": {
      "post": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Telemetry",
        "operationId": "telemetry_api_telemetry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/gis": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Gis",
        "operationId": "gis_api_gis_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/assistant": {
      "post": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Assistant",
        "operationId": "assistant_api_assistant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Question"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/gis/nearby": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Nearby",
        "operationId": "nearby_api_gis_nearby_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude"
            }
          },
          {
            "name": "radius_m",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 50000,
              "exclusiveMinimum": 0,
              "default": 1000,
              "title": "Radius M"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Audit Logs",
        "operationId": "audit_logs_api_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/verify": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Check Chain",
        "operationId": "check_chain_api_audit_verify_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/notifications": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Notifications",
        "operationId": "notifications_api_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Mark Read",
        "operationId": "mark_read_api_notifications__notification_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/cases.csv": {
      "get": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Export Cases",
        "operationId": "export_cases_api_reports_cases_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Question": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 5,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "question"
        ],
        "title": "Question"
      },
      "SensorInput": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "source": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Source"
          },
          "event_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 5,
            "title": "Event Id"
          },
          "metric": {
            "type": "string",
            "enum": [
              "methane",
              "temperature",
              "dust",
              "vibration"
            ],
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Unit"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "asset_id",
          "source",
          "event_id",
          "metric",
          "value",
          "unit",
          "observed_at"
        ],
        "title": "SensorInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}